backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deduptickets"
version = "0.1.0"
description = "Ticket Deduplication & Clustering Dashboard"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "DedupTickets Team" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "azure-cosmos>=4.5.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
    "httpx>=0.26.0",
    "bandit[toml]>=1.7.0",
]

[project.scripts]
deduptickets = "deduptickets.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/deduptickets"]

# ============================================================================
# RUFF CONFIGURATION (Constitution Principle VI)
# ============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
    "S",      # flake8-bandit (security)
    "ASYNC",  # flake8-async
]
ignore = [
    "PLR0913",  # Too many arguments - sometimes necessary for Pydantic models
    "PLR2004",  # Magic value used in comparison
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",     # Allow assert in tests
    "ARG001",   # Allow unused function arguments (fixtures)
    "PLR2004",  # Allow magic values in tests
]
"src/deduptickets/models/*.py" = [
    "TC003",    # datetime is used at runtime in Field defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ============================================================================
# PYTEST CONFIGURATION (Constitution Principle III)
# ============================================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src/deduptickets",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests (fast, no I/O)",
    "integration: Integration tests (require Cosmos DB Emulator)",
    "contract: API contract tests",
    "slow: Slow tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# ============================================================================
# COVERAGE CONFIGURATION
# ============================================================================
[tool.coverage.run]
source = ["src/deduptickets"]
branch = true
omit = [
    "*/__init__.py",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true

# ============================================================================
# BANDIT CONFIGURATION (Security scanning)
# ============================================================================
[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # Allow assert in production code for Pydantic validators
